Add TXT records

+++ Sauron/BackEnd.pm
@@ -1954,11 +1954,14 @@ sub get_host($$) {
 		  "Priority,Weight,Port,Target,Comments",
 		  "type=1 AND ref=$id ORDER BY port,pri,weight",$rec,'srv_l');
   get_array_field("sshfp_entries",6,"id,algorithm,hashtype,fingerprint,comment",
-		  "Algorithm,Type,Fingerprint",
+		  "Algorithm,Type,Fingerprint,Comments",
 		  "type=1 AND ref=$id ORDER BY algorithm,hashtype,fingerprint",$rec,'sshfp_l');
   get_array_field("tlsa_entries",6,"id,usage,selector,matching_type,association_data,comment",
 		  "Usage,Selector,Matching Type,Asociation Data,Comments",
 		  "type=1 AND ref=$id ORDER BY usage,selector,matching_type,association_data",$rec,'tlsa_l');
+  get_array_field("txt_entries",3,"id,txt,comment",
+		  "Text,Comments",
+		  "type=2 AND ref=$id ORDER BY txt",$rec,'txt_l');
 # Get CNAME aliases.
   get_array_field("hosts",5,"0,id,domain,type,1","Domain,cname",
 	          "type=4 AND alias=$id ORDER BY domain",$rec,'alias_l');
@@ -2141,23 +2144,27 @@ sub update_host($) {
 			"usage,selector,matching_type,association_data,comment,type,ref",
 			'tlsa_l',$rec,"1,$id");
   if ($r < 0) { db_rollback(); return -20; }
+  $r=update_array_field("txt_entries",6,
+			"txt,comment,type,ref",
+			'txt_l',$rec,"1,$id");
+  if ($r < 0) { db_rollback(); return -21; }
   $r=update_array_field("a_entries",4,"ip,reverse,forward,host",
 			'ip',$rec,"$id");
-  if ($r < 0) { db_rollback(); return -21; }
+  if ($r < 0) { db_rollback(); return -22; }
 
   if ($rec->{type}==7) {
     $r=update_array_field("arec_entries",2,"arec,host",
 			  'alias_a',$rec,"$id");
-    if ($r < 0) { db_rollback(); return -22; }
+    if ($r < 0) { db_rollback(); return -23; }
   }
 
   $r=update_array_field("group_entries",2,"grp,host",
 			'subgroups',$rec,"$id");
-  if ($r < 0) { db_rollback(); return -23; }
+  if ($r < 0) { db_rollback(); return -24; }
 
   $r=update_array_field("dhcp_entries",3,"dhcp,comment,type,ref",
 			'dhcp_l6',$rec,"13,$id");
-  if ($r < 0) { db_rollback(); return -24; }
+  if ($r < 0) { db_rollback(); return -25; }
 
   return db_commit();
 }
@@ -2242,6 +2249,10 @@ sub delete_host($) {
   # tlsa_entries
   $res=db_exec("DELETE FROM tlsa_entries WHERE type=1 AND ref=$id;");
   if ($res < 0) { db_rollback(); return -16; }
+
+  # txt_entries
+  $res=db_exec("DELETE FROM txt_entries WHERE type=2 AND ref=$id;");
+  if ($res < 0) { db_rollback(); return -17; }
 
   # group_entries
   $res=db_exec("DELETE FROM group_entries WHERE host=$id;");
@@ -2341,6 +2352,11 @@ sub add_host($) {
 			 'tlsa_l',$rec,'type,ref',"1,$id");
   if ($res < 0) { db_rollback(); return -10; }
 
+  # TXTs
+  $res = add_array_field('txt_entries','txt,comment',
+			 'txt_l',$rec,'type,ref',"2,$id");
+  if ($res < 0) { db_rollback(); return -10; }
+
   # ARECs
   if ($rec->{type}==7) {
     $res=db_exec("INSERT INTO arec_entries (host,arec) VALUES($id,$a_id);");
@@ -2363,7 +2379,7 @@ sub get_host_types() {
     return (0 => 'Any type', 1 => 'Host', 2 => 'Delegation', 3 => 'Plain MX',
 	    4 => 'Alias', 5 => 'Printer', 6 => 'Glue', 7 => 'AREC Alias',
 	    8 => 'SRV', 9 => 'DHCP only', 10 => 'Zone', 11=>'SSHFP only',
-            12 => 'TLSA only',
+            12 => 'TLSA only', 13 => 'TXT',
 	    101 => 'Host reservation');
 }
